Share Spring context across sbs3 scheduled-task tests
ScheduledTaskServiceTest and ScheduledTasksRegistryTest each bootstrapped
their own Spring ApplicationContext because each imported a different inner
@TestConfiguration, giving them distinct MergedContextConfiguration cache
keys.

Introduce AbstractScheduledTasksIntegrationTest, an abstract base that
carries the shared @SpringBootTest + @Import of a single superset
@TestConfiguration. Both tests now extend it, so their inherited context
configuration is identical and Spring reuses one cached context (confirmed
via the spring-test-profiler report: both classes map to one MOCK context).

ScheduledTasksRegistryTest is read-only and its assertions are made state-
and order-independent (contains the expected task ids / has at least one of
each task type) so the context mutated by ScheduledTaskServiceTest cannot
cause order-dependent failures.

AxelixScheduledTasksEndpointTest is intentionally left on its own context
(RANDOM_PORT web environment + property source + JWT auth configuration).
